requirements.txt
Django
gunicorn
djangorestframework
djangochannelsrestframework
djangorestframework-simplejwt[crypto]
channels-redis
channels["daphne"]